Fix type of variable to hold vo_paramNext() return value
`vo_paramNext()` returns an `int`, and specifically returns a negative value
if there are arguments left after processing the options.

Depending on the signedness of `char`, this may be converted to a positive
value if the return value is converted to a `char`. This happens f.e. on ARM
processors, where the char is `unsigned char` by default. Since the option
parsing is stopped when this is `<=0`, additional (non-option) arguments lead
to an errornous processing.

The correct type to hold the return value is `int`, and this was already set
in some VO applications (f.e. `votjoin`). This patch changes this for the
remaining apps.

This fixes the read problem of VOTables on arm processors.

This was already proposed as pull request
[iraf/iraf-v216#112](https://iraf-community.github.io/iraf-v216/issues/112)
but got ignored at that time.